requirements.txt
icloudpd==1.25.0
PyYAML==6.0.2
